+++ apps/docs/src/content/03-patterns/02-codesnippets/multi-upload/examples/default.tsx
@@ -2,8 +2,10 @@ import {
   Button,
   FileCard,
   FileCardList,
+  FileDropZone,
   FileField,
-  Label,
+  Heading,
+  IconUpload,
   Section,
 } from "@mittwald/flow-react-components";
 import { useState } from "react";
@@ -13,19 +15,23 @@ export default () => {
 
   return (
     <Section>
-      <FileField
+      <FileDropZone
         multiple
-        onChange={(value) => {
-          if (value) {
-            setFiles([...files, ...value]);
+        onChange={(droppedFiles) => {
+          if (droppedFiles) {
+            setFiles((existingFiles) => [
+              ...existingFiles,
+              ...droppedFiles,
+            ]);
           }
         }}
       >
-        <Label>Dateien</Label>
-        <Button variant="outline" color="secondary">
-          Auswählen
-        </Button>
-      </FileField>
+        <IconUpload />
+        <Heading>Dateien ablegen</Heading>
+        <FileField>
+          <Button>Dateien auswählen</Button>
+        </FileField>
+      </FileDropZone>
 
       <FileCardList>
         {files.map((file) => (